use size_t for buffer sizes (silences several warnings)

+++ utils/skygw_utils.h
@@ -126,8 +126,8 @@ int               skygw_thread_start(skygw_thread_t* thr);
 skygw_thr_state_t skygw_thread_get_state(skygw_thread_t* thr);
 pthread_t         skygw_thread_gettid(skygw_thread_t* thr);
 
-int get_timestamp_len(void);
-int snprint_timestamp(char* p_ts, int tslen);
+size_t get_timestamp_len(void);
+int snprint_timestamp(char* p_ts, size_t tslen);
 
 EXTERN_C_BLOCK_BEGIN
